+++ jest.config.js
@@ -11,7 +11,6 @@ module.exports = {
   testPathIgnorePatterns: [
     '/node_modules/',
     '/lib/',
-    'create-cli-utils/',
   ],
   preset: 'ts-jest',
 };
