+++ packages/babel-plugin-transform-jsx-stylesheet/src/__tests__/module.ts
@@ -1,8 +1,8 @@
-import syntaxJSX from 'babel-plugin-syntax-jsx';
+import * as syntaxJSX from 'babel-plugin-syntax-jsx';
 import { transform } from 'babel-core';
 import jSXStylePlugin from '../index';
 
-export default function getTransformCode(code, opts) {
+export default function getTransformCode(code, opts?) {
   return transform(code, {
     plugins: [
       [jSXStylePlugin, opts],
